+++ spec/features/users/users_can_manage_user_accounts_spec.rb
@@ -77,6 +77,28 @@ RSpec.feature "Users can manage user accounts" do
     expect(page).to have_content("There is a problem")
   end
 
+  scenario "existing users can be deactivated" do
+    existing_user = create(:user)
+
+    visit edit_user_path(existing_user)
+    uncheck "Active"
+
+    click_on "Save user"
+
+    expect(existing_user.reload.active).to be false
+  end
+
+  scenario "inactive users can be activated" do
+    existing_user = create(:inactive_user)
+
+    visit edit_user_path(existing_user)
+    check "Active"
+
+    click_on "Save user"
+
+    expect(existing_user.reload.active).to be true
+  end
+
   context "then the users team is nil becuase it is a legacy account" do
     scenario "no team is shown" do
       other_user = User.new(