+++ spec/models/user_spec.rb
@@ -261,6 +261,44 @@ RSpec.describe User do
     end
   end
 
+  describe "#active" do
+    it "returns true when deactivated_at is nil" do
+      user = build(:user)
+
+      expect(user.active).to be true
+    end
+
+    it "returns false when deactivated_at has a value" do
+      user = build(:inactive_user)
+
+      expect(user.active).to be false
+    end
+
+    it "can be set with 1 and 0 strings" do
+      user = build(:user)
+
+      user.update!(active: "0")
+
+      expect(user.active).to be false
+
+      user.update!(active: "1")
+
+      expect(user.active).to be true
+    end
+
+    it "can be set with true and false" do
+      user = build(:user)
+
+      user.update!(active: false)
+
+      expect(user.active).to be false
+
+      user.update!(active: true)
+
+      expect(user.active).to be true
+    end
+  end
+
   def valid_user_attributes
     {
       first_name: "First",